# Conflicts:
#	.idea/workspace.xml
#	pom.xml
#	src/main/java/org/example/dao/Client.java
#	src/main/java/org/example/dao/Owner.java
#	src/main/java/org/example/dao/Person.java
#	src/main/java/org/example/dao/Role.java
#	src/main/java/org/example/security/WebSecurityConfig.java
#	src/main/resources/application.properties

+++ src/main/java/org/example/service/Impl/OrderServiceImpl.java
@@ -1,20 +1,25 @@
 package org.example.service.Impl;
 
 
-import org.example.dao.Client;
 import org.example.dao.Order;
+import org.example.dao.Person;
 import org.example.dao.Station;
 import org.example.repository.OrderRepository;
 import org.example.service.OrderService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import javax.transaction.Transactional;
 import java.util.List;
 @Service
+@Transactional
 public class OrderServiceImpl implements OrderService {
 
+    private final OrderRepository orderRepository;
     @Autowired
-    private OrderRepository orderRepository;
+    public OrderServiceImpl(OrderRepository orderRepository) {
+        this.orderRepository = orderRepository;
+    }
 
     @Override
     public Order createOrder(Order order) {
@@ -28,7 +33,9 @@ public class OrderServiceImpl implements OrderService {
 
     @Override
     public void deleteOrder(long orderId) {
-        orderRepository.deleteById(orderId);
+        if(orderRepository.existsById(orderId)) {
+            orderRepository.deleteById(orderId);
+        }
     }
 
     @Override
@@ -42,12 +49,18 @@ public class OrderServiceImpl implements OrderService {
     }
 
     @Override
-    public Client getClientByOrderId(long orderId) {
-        return orderRepository.findById(orderId).get().getClient();
+    public Person getPersonByOrderId(long orderId) {
+        if(orderRepository.existsById(orderId)) {
+            return orderRepository.findById(orderId).get().getPerson();
+        }
+        return null;
     }
 
     @Override
     public Station getStationByOrderId(long orderId) {
-        return orderRepository.findStationById(orderId).orElse(null);
+        if(orderRepository.existsById(orderId)) {
+            return orderRepository.findById(orderId).get().getStation();
+        }
+        return null;
     }
 }
