+++ src/main/java/org/example/service/Impl/StationServiceServiceImpl.java
@@ -7,12 +7,18 @@ import org.example.service.StationServiceService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import javax.transaction.Transactional;
 import java.util.List;
 @Service
+@Transactional
 public class StationServiceServiceImpl implements StationServiceService {
 
+
+    private final StationServiceRepository stationServiceRepository;
     @Autowired
-    private StationServiceRepository stationServiceRepository;
+    public StationServiceServiceImpl(StationServiceRepository stationServiceRepository) {
+        this.stationServiceRepository = stationServiceRepository;
+    }
 
     @Override
     public StationService getStationServiceById(long id) {
@@ -26,17 +32,25 @@ public class StationServiceServiceImpl implements StationServiceService {
 
     @Override
     public StationService createStationService(StationService stationService) {
+        if (stationServiceRepository.existsById(stationService.getId())) {
+            return stationServiceRepository.findById(stationService.getId()).get();
+        }
         return stationServiceRepository.save(stationService);
     }
 
     @Override
     public StationService updateStationService(StationService stationService) {
-        return stationServiceRepository.save(stationService);
+        if (stationServiceRepository.existsById(stationService.getId())) {
+            return stationServiceRepository.save(stationService);
+        }
+        return null;
     }
 
     @Override
     public void deleteStationServiceById(long id) {
-        stationServiceRepository.deleteById(id);
+        if (stationServiceRepository.existsById(id)) {
+            stationServiceRepository.deleteById(id);
+        }
     }
 
 }
