+++ src/main/java/org/example/service/Impl/UserServiceImpl.java
@@ -1,41 +1,70 @@
 package org.example.service.Impl;
 
-
-import org.example.dao.Client;
-import org.example.dao.Person;
-import org.example.dao.Role;
-import org.example.enums.PersonType;
+import org.example.dao.*;
 import org.example.repository.PersonRepository;
 import org.example.service.UserService;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.userdetails.*;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
+import javax.transaction.Transactional;
 import java.util.List;
 
 @Service
-
+@Transactional
 public class UserServiceImpl implements UserDetailsService, UserService {
+
+    private static final Logger log = LoggerFactory.getLogger(UserServiceImpl.class);
+    private final PersonRepository personRepository;
+
     @Autowired
-    private PersonRepository personRepository;
+    private PasswordEncoder passwordEncoder;
+
+
     @Override
     public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        return personRepository.findByLogin(username).orElseThrow(() -> new UsernameNotFoundException(username));
+        return personRepository.findPersonEntityByLogin(username).orElseThrow(()->new UsernameNotFoundException(username));
     }
+
+    @Autowired
+    public UserServiceImpl(PersonRepository personRepository) {
+        this.personRepository = personRepository;
+    }
+
 
     @Override
     public Person createUser(Person person) {
-        return personRepository.save(person);
+        Person personByEmail = personRepository.findByEmail(person.getEmail()).orElse(null);
+        if (personByEmail==null) {
+            String encriptPass = passwordEncoder.encode(person.getPassword());
+            person.setPassword(encriptPass);
+            return personRepository.save(person);
+        }
+        log.error("Пользователь с таким email уже существует");
+        return personByEmail;
     }
 
     @Override
     public Person updateUser(Person person) {
+        if (personRepository.findByLogin(person.getLogin()).isPresent() || personRepository.findByEmail(person.getEmail()).isPresent()) {
+            return null;
+        }
         return personRepository.save(person);
     }
 
     @Override
     public void deleteUserById(Long id) {
-        personRepository.deleteById(id);
+        if (!personRepository.findById(id).isPresent()) {
+            log.info("not deleting user with id: " + id);
+            throw new UsernameNotFoundException("User not found");
+        } else {
+            personRepository.deleteById(id);
+            log.info("deleting user with id: " + id);
+        }
+
     }
 
     @Override
@@ -45,7 +74,7 @@ public class UserServiceImpl implements UserDetailsService, UserService {
 
     @Override
     public Person getUserById(long id) {
-        return  personRepository.findById(id).orElse(null);
+        return personRepository.findById(id).orElse(null);
     }
 
     @Override
@@ -55,22 +84,6 @@ public class UserServiceImpl implements UserDetailsService, UserService {
 
     @Override
     public List<Person> getAllUsers() {
-
         return personRepository.findAll();
-    }
-
-    @Override
-    public List<Role> getRolesByUserId(long id) {
-        return personRepository.findRolesById(id);
-    }
-
-    @Override
-    public List<PersonType> getPersonTypesByPersonId(long id) {
-        return personRepository.findPersonTypesById(id);
-    }
-
-    @Override
-    public List<Client> getClientsById(long id) {
-        return List.of();
     }
 }
